--- app/src/main/java/com/example/pace/ui/add_schedule/search_box/MyLocationFragment.kt
@@ -1,4 +1,0 @@
-package com.example.pace.ui.add_schedule.search_box
-
-class MyLocationFragment {
-}
