+++ app/src/main/java/com/example/pace/ui/search_box/MapFragment.kt
@@ -1,0 +1,4 @@
+package com.example.pace.ui.search_box
+
+class MapFragment {
+}
