+++ app/src/main/java/com/example/pace/ui/search_box/MyLocationFragment.kt
@@ -1,0 +1,4 @@
+package com.example.pace.ui.search_box
+
+class MyLocationFragment {
+}
