+++ app/src/main/java/com/example/pace/ui/search_box/SearchLocationActivity.kt
@@ -1,0 +1,4 @@
+package com.example.pace.ui.search_box
+
+class SearchLocationActivity {
+}
