+++ app/src/main/res/drawable/home_no_item.xml
@@ -1,0 +1,18 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="116dp"
+    android:height="121dp"
+    android:viewportWidth="116"
+    android:viewportHeight="121">
+  <path
+      android:pathData="M54.55,27.27C54.55,12.22 42.33,0 27.27,0C12.22,0 -0,12.22 -0,27.27C-0,41.92 11.53,53.85 25.99,54.52C26.67,54.55 27.22,54 27.22,53.32L27.22,40.12C27.22,39.45 26.67,38.91 26,38.84C20.17,38.21 15.64,33.26 15.64,27.27C15.64,20.87 20.84,15.61 27.3,15.61C33.76,15.61 38.82,20.73 38.93,27.08L38.93,54.55L54.58,54.55L54.58,27.27L54.55,27.27Z"
+      android:fillColor="#D9D9D9"/>
+  <path
+      android:pathData="M61.03,92.74C61.03,107.8 73.24,120.01 88.3,120.01C103.36,120.01 115.57,107.8 115.57,92.74C115.57,78.1 104.04,66.16 89.58,65.5C88.91,65.46 88.36,66.01 88.36,66.69L88.36,79.89C88.36,80.56 88.9,81.1 89.57,81.17C95.4,81.81 99.93,86.75 99.93,92.74C99.93,99.14 94.73,104.4 88.27,104.4C81.82,104.4 76.75,99.28 76.64,92.93L76.64,65.47L61,65.47L61,92.74L61.03,92.74Z"
+      android:fillColor="#D9D9D9"/>
+  <path
+      android:pathData="M27.27,120.01C41.89,120.01 53.85,108.48 54.52,93.99C54.55,93.32 54,92.77 53.32,92.77L40.12,92.77C39.45,92.77 38.91,93.31 38.84,93.98C38.21,99.81 33.26,104.35 27.27,104.35C20.87,104.35 15.61,99.14 15.61,92.68C15.61,86.23 20.73,81.16 27.08,81.05L53.32,81.05C54,81.05 54.55,80.5 54.55,79.83L54.55,54L39.05,54L39.05,65.44L27.27,65.44C12.22,65.44 -0,77.66 -0,92.71C-0,107.77 12.19,120.01 27.27,120.01Z"
+      android:fillColor="#D9D9D9"/>
+  <path
+      android:pathData="M88.3,0C73.68,0 61.72,11.53 61.06,26.02C61.03,26.7 61.58,27.25 62.25,27.25L75.45,27.25C76.12,27.25 76.66,26.7 76.74,26.03C77.37,20.2 82.31,15.67 88.3,15.67C94.7,15.67 99.96,20.87 99.96,27.33C99.96,33.79 94.84,38.85 88.5,38.96L62.25,38.96C61.57,38.96 61.03,39.51 61.03,40.19L61.03,66.01L76.53,66.01L76.53,54.58L88.3,54.58C103.36,54.58 115.57,42.36 115.57,27.3C115.57,12.25 103.39,0 88.3,0Z"
+      android:fillColor="#D9D9D9"/>
+</vector>
