+++ app/src/main/res/layout/fragment_home.xml
@@ -1,12 +1,39 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:background=""
     xmlns:app="http://schemas.android.com/apk/res-auto">
-    <TextView
+    <LinearLayout
+        android:id="@+id/home_background"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:text="home"
-        android:gravity="center"/>
+        android:gravity="center"
+        android:orientation="vertical">
+        <ImageView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:src="@drawable/home_no_item"/>
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="오늘 나의 페이스"
+            android:textAppearance="@style/TextAppearance.App.BodySm.Medium"
+            android:layout_marginTop="12dp"/>
+    </LinearLayout>
+    <LinearLayout
+        android:id="@+id/home_add_schedule_ll"
+        android:layout_width="match_parent"
+        android:layout_height="48dp"
+        android:layout_marginStart="20dp"
+        android:layout_marginEnd="20dp"
+        android:layout_gravity=""
+        android:background="@drawable/box_schedule">
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="일정 추가"
+            android:textAppearance="@style/TextAppearance.App.CaptionLg.Semibold"/>
+    </LinearLayout>
 
-</androidx.constraintlayout.widget.ConstraintLayout>
+</FrameLayout>
